added flag to visualizer to let the policy be run for shorter or longer than max_path_length

+++ cistar-dev/cistar/visualizer_CISTAR.py
@@ -23,7 +23,7 @@ if __name__ == "__main__":
                         help='Prefix for all generated plots')
     parser.add_argument('--use_sumogui', type=bool, default=True,
                         help='Flag for using sumo-gui vs sumo binary')
-    parser.add_argument('--run_long', type=int, default=1,
+    parser.add_argument('--run_long', type=float, default=1,
                         help='Number by which to increase max_path_length')
     args = parser.parse_args()
